requirements.txt
pysoem
fastapi
uvicorn[standard]
websockets
pyyaml
